Device initialization file implementation (part 1).
Added UI for external device configuration.
Instrument mapping portion of the device intialization file is now working.
Fixed a few spots where return value of getChannelPreset() is ignored.
Fixed layout of the preset selection dialog.

+++ core/qmpmidioutfluid.cpp
@@ -123,7 +123,7 @@ std::vector<std::pair<uint16_t,std::string>> qmpMidiOutFluid::getBankList()
 {
 	return bnk;
 }
-std::vector<std::pair<uint8_t,std::string>> qmpMidiOutFluid::getPresets(int bank)
+std::vector<std::pair<uint8_t,std::string>> qmpMidiOutFluid::getPresets(uint16_t bank)
 {
 	std::vector<std::pair<uint8_t,std::string>> ret;
 	if(pst.find(bank)==pst.end())return ret;
@@ -137,6 +137,7 @@ std::vector<std::pair<uint8_t,std::string>> qmpMidiOutFluid::getPresets(int bank
 std::string qmpMidiOutFluid::getPresetName(uint16_t bank,uint8_t preset)
 {
 	if(pst.find(bank)==pst.end())return "";
+	//should consult fluidsynth instead? (4 bank mapping methods)
 	return pst[bank][preset];
 }
 bool qmpMidiOutFluid::getChannelPreset(int ch,uint16_t *bank,uint8_t *preset,std::string &presetname)
