+++ core/qmpmidioutfluid.hpp
@@ -41,7 +41,7 @@ class qmpMidiOutFluid:public qmpMidiOutDevice,public IFluidSettings
 		void onMapped(uint8_t ch,int refcnt);
 		void onUnmapped(uint8_t ch,int refcnt);
 		std::vector<std::pair<uint16_t,std::string>> getBankList();
-		std::vector<std::pair<uint8_t,std::string>> getPresets(int bank);
+		std::vector<std::pair<uint8_t,std::string>> getPresets(uint16_t bank);
 		std::string getPresetName(uint16_t bank,uint8_t preset);
 		bool getChannelPreset(int ch,uint16_t *bank,uint8_t *preset,std::string &presetname);
 		uint8_t getInitialCCValue(uint8_t cc);
